# server/swagger/document.yml
swagger: "2.0"
info:
  description: "EPIC-MAIL WEB APP ON HEROKU"
  version: "1.0.0"
  title: "EPIC-MAIL"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "rukundoemma@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:4000"
basePath: "/api/v1"
...
schemes:
- "http"
- "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWT_Token_User:
    type: apiKey
    in: header
    name: epicmail authorasation  
paths:
...
 /messages:
    post:
      summary: "Send Message"
      tags:
        - "message"
      operationId: "sendMessage" 
      produces:
      - application/json
      parameters:
      - name: "id"
        in: "path"
        description: "Message id"
        required: true
        type: "integer"
        format: "int5"
      - name: "subject"
        in: "path"
        description: "Subject"
        required: true
        type: "string"
        format: "int5"  
      - name: "message"
        in: "path"
        description: "Message"
        required: true
        type: "string"
        format: "int5" 
      - name: "senderId"
        in: "path"
        description: "Sender id"
        required: true
        type: "integer"
        format: "int5"  
      - name: "receiverId"
        in: "path"
        description: "Receiver id"
        required: true
        type: "integer"
        format: "int5"  
      - name: "parentMessageId"
        in: "path"
        description: "This is Replied message id"
        required: true
        type: "integer"
        format: "int5"
      - name: "status"
        in: "path"
        description: "Messages status, this may be read, sent or draft "
        required: true
        type: "string"
      - name: "createdOn"
        in: "path"
        description: "Timestamp of the messages"
        required: true
        type: "string"
        format: "2019-03-04T05:44:00.494Z"
      responses:
        '200':
          description: Successful operation.
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 200
              data:
                type: object
                $ref: "#/definitions/message"            

           
    get:
      summary: "Get all received messages"
      tags:
        - "message"
      operationId: "getAllReceivedMessages" 
      produces:
      - application/json
      responses:
        '200':
          description: Successful operation:This will return  All user's received messages. 
                      NOTE:All  received message are all of those inbox messages which their receiverId is equal to user's id. this will avoid spaming. 
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 200
              data:
                type: array
                items:
                    $ref: "#/definitions/message"  
        '404':
          description: No any received message found
          schema: 
            type: object
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: You have not any received message!!!
...
definitions:
...
  message:
    type: object
    properties:
      id:
        type: integer
        example: 2
      createOn:
        type: string
        example:  "2019-03-04T05:44:00.494Z"
      subject:
        type: string
        example: Invitation to mariage
      senderId:
        type: integer
        example: 3
      receiverId:
        type: integer
        example: 2
      message: 
        type: string
        example: Hello guys, we would to invite you to our mariage celemony which will take place at kacyiro suday pack 11:00 am, Thank you!
      parentMessageId:
        type: integer
        example: 1
      status:
        type: string
        example: sent 
...